+++ Plugins/ActorPools/Source/NexusActorPools/Private/NActorPoolListViewEntry.cpp
@@ -1,0 +1,11 @@
+﻿#include "NActorPoolListViewEntry.h"
+
+void UNActorPoolListViewEntry::NativeDestruct()
+{
+	Super::NativeDestruct();
+}
+
+void UNActorPoolListViewEntry::NativeOnListItemObjectSet(UObject* ListItemObject)
+{
+	INListViewEntry::NativeOnListItemObjectSet(ListItemObject);
+}