+++ Plugins/ActorPools/Source/NexusActorPools/Private/NActorPoolSubsystem.cpp
@@ -7,6 +7,16 @@
 #include "NActorPoolSet.h"
 #include "NActorPoolSpawnerComponent.h"
 
+namespace NEXUS::ActorPools
+{
+	static bool bTrackStats = false;
+	static FAutoConsoleVariableRef CVAR_TrackStats(
+		TEXT("NEXUS.ActorPools.TrackStats"),
+		bTrackStats,
+		TEXT("Track stats via internal stats system."),
+		ECVF_Default | ECVF_Preview);
+}
+
 void UNActorPoolSubsystem::Deinitialize()
 {
 	for (TPair<UClass*, TUniquePtr<FNActorPool>>& Pool : ActorPools)
@@ -22,12 +32,13 @@ void UNActorPoolSubsystem::Initialize(FSubsystemCollectionBase& Collection)
 	Super::Initialize(Collection);
 	
 	const UNActorPoolsSettings* Settings = UNActorPoolsSettings::Get();
-	bStatsEnabled = Settings->bTrackStats;
+	
+	UnknownBehaviour = Settings->UnknownBehaviour;
 	
 	// Check if we have anything to automatically load
 	if (Settings->AlwaysCreateSets.Num() == 0) return;
 	
-	// Check if we have any ignores
+	// Check if we have any prefixes to ignore
 	if (Settings->IgnoreWorldPrefixes.Num() > 0)
 	{
 		bool bShouldIgnoreLevel = false;
@@ -51,14 +62,15 @@ void UNActorPoolSubsystem::Initialize(FSubsystemCollectionBase& Collection)
 
 bool UNActorPoolSubsystem::IsTickable() const
 {
-	return bStatsEnabled || bHasTickableActorPools || bHasTickableSpawners;
+	
+	return NEXUS::ActorPools::bTrackStats || bHasTickableActorPools || bHasTickableSpawners;
 }
 
 void UNActorPoolSubsystem::Tick(float DeltaTime)
 {
-	CONDITIONAL_SCOPE_CYCLE_COUNTER(STAT_ActorPoolSubsystemTick, bStatsEnabled);
+	CONDITIONAL_SCOPE_CYCLE_COUNTER(STAT_ActorPoolSubsystemTick, NEXUS::ActorPools::bTrackStats);
 	
-	if (bStatsEnabled)
+	if (NEXUS::ActorPools::bTrackStats)
 	{
 		SET_DWORD_STAT(STAT_ActorPoolCount, ActorPools.Num())
 		SET_DWORD_STAT(STAT_InActors, 0)
@@ -81,7 +93,7 @@ void UNActorPoolSubsystem::Tick(float DeltaTime)
 		}
 	}
 
-	if (bStatsEnabled)
+	if (NEXUS::ActorPools::bTrackStats)
 	{
 		for ( auto Pair = ActorPools.CreateConstIterator(); Pair; ++Pair )
 		{
@@ -101,8 +113,9 @@ bool UNActorPoolSubsystem::CreateActorPool(TSubclassOf<AActor> ActorClass, const
 	
 	if (!ActorPools.Contains(ActorClass))
 	{
-		ActorPools.Add(ActorClass, MakeUnique<FNActorPool>(GetWorld(), ActorClass, Settings));
+		const TUniquePtr<FNActorPool>& NewPool = ActorPools.Add(ActorClass, MakeUnique<FNActorPool>(GetWorld(), ActorClass, Settings));
 		UE_LOG(LogNexusActorPools, Verbose, TEXT("Creating a new FNActorPool(%s) in UWorld(%s), raising the total pool count to %i."), *ActorClass->GetName(), *GetWorld()->GetName(), ActorPools.Num());
+		OnActorPoolAdded.Broadcast(NewPool.Get());
 		return true;
 	}
 	return false;
@@ -182,17 +195,38 @@ void UNActorPoolSubsystem::SpawnActor(TSubclassOf<AActor> ActorClass, FVector Po
 
 bool UNActorPoolSubsystem::ReturnActor(AActor* Actor)
 {
-	const UClass* ActorClass = Actor->GetClass();
+	UClass* ActorClass = Actor->GetClass();
 	if (ActorPools.Contains(ActorClass))
 	{
 		return (*ActorPools.Find(ActorClass))->Return(Actor);
 	}
-	if (bDestroyUnknownReturnedActors)
+	
+	switch (UnknownBehaviour)
 	{
-		UE_LOG(LogNexusActorPools, Verbose, TEXT("Attempting to destroy an unknown AActor(%s)) set to the UNActorPoolSubsystem."), *Actor->GetPathName());
-		return Actor->Destroy();
+		using enum ENActorPoolUnknownBehaviour;
+		case CreateDefaultPool:
+		{
+			UE_LOG(LogNexusActorPools, Log, TEXT("Creating a new pool via ReturnActor for %s (%s), raising the total pool count to %i."),
+				*ActorClass->GetName(), *GetWorld()->GetName(), ActorPools.Num());
+			const TUniquePtr<FNActorPool>& NewPool = ActorPools.Add(ActorClass, MakeUnique<FNActorPool>(GetWorld(), ActorClass));
+			OnActorPoolAdded.Broadcast(NewPool.Get());
+			if (NewPool->DoesSupportInterface())
+			{
+				INActorPoolItem* ActorItem = Cast<INActorPoolItem>(Actor);
+				ActorItem->InitializeActorPoolItem(NewPool.Get());		
+			}
+			return NewPool->Return(Actor);
+		}
+		case Ignore:
+			return false;
+		case Destroy: 
+		default:
+			if (Actor->IsRooted())
+			{
+				return false;
+			}
+			return Actor->Destroy();
 	}
-	return false;
 }
 
 void UNActorPoolSubsystem::RegisterTickableSpawner(UNActorPoolSpawnerComponent* TargetComponent)
@@ -205,9 +239,4 @@ void UNActorPoolSubsystem::UnregisterTickableSpawner(UNActorPoolSpawnerComponent
 {
 	TickableSpawners.Remove(TargetComponent);
 	bHasTickableSpawners = (TickableSpawners.Num() > 0);
-}
-
-void UNActorPoolSubsystem::SetDestroyUnknownReturnedActors(const bool ShouldDestroy)
-{
-	bDestroyUnknownReturnedActors = ShouldDestroy;
 }